02.agents/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pydantic>=2.7.0
openai>=1.30.0
psycopg2-binary>=2.9.9
qdrant-client>=1.9.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
